+++ docs/assets/UluConditionalText.stories-2Bu-d9p3.js
@@ -1,0 +1,21 @@
+import{ap as n}from"./iframe-BAda3xFs.js";import"./preload-helper-BJwshlQW.js";const l={component:n,tags:["autodocs"]},e={args:{text:"This text is set"}},t={args:{text:0}},r={args:{text:null}},s={args:{text:void 0}};class c{constructor(o){this.text=o}toString(){return this.text}}const a={args:{text:new c("This is a class with toString")}};e.parameters={...e.parameters,docs:{...e.parameters?.docs,source:{originalSource:`{
+  args: {
+    text: "This text is set"
+  }
+}`,...e.parameters?.docs?.source}}};t.parameters={...t.parameters,docs:{...t.parameters?.docs,source:{originalSource:`{
+  args: {
+    text: 0
+  }
+}`,...t.parameters?.docs?.source}}};r.parameters={...r.parameters,docs:{...r.parameters?.docs,source:{originalSource:`{
+  args: {
+    text: null
+  }
+}`,...r.parameters?.docs?.source}}};s.parameters={...s.parameters,docs:{...s.parameters?.docs,source:{originalSource:`{
+  args: {
+    text: undefined
+  }
+}`,...s.parameters?.docs?.source}}};a.parameters={...a.parameters,docs:{...a.parameters?.docs,source:{originalSource:`{
+  args: {
+    text: new toStringObject("This is a class with toString")
+  }
+}`,...a.parameters?.docs?.source}}};const m=["StringValue","NumberValue","NullValue","UndefinedValue","ObjectWithToString"];export{r as NullValue,t as NumberValue,a as ObjectWithToString,e as StringValue,s as UndefinedValue,m as __namedExportsOrder,l as default};