+++ lib/components/forms/UluFormCheckbox.vue
@@ -8,8 +8,7 @@
   >
   <label :for="id">
     <slot>
-      {{ label }}
-      <UluFormRequiredChar v-if="required" />
+      {{ label }}<UluFormRequiredChar v-if="required" />
     </slot>
   </label>
 </template>